HBASE-30129 Split TestFromClientSide3 (#8163)
(cherry picked from commit fab7c6ed7cb93f488a717af87f677fbeceb5369e)
